Use `to_unsafe_h` to get a hash of the current params
The `with_indifferent_access` is called when we try to consolidate our
changes back with `ActionController::Parameters.new` which has been
deprecated for use ahead of being removed in Rails 5.1.

Calling `to_unsafe_h` instead of `clone` fixes the problem by providing
us with a hash representation of the parameters that responds to
`with_indifferent_access`. (Deliberately invoking the "unsafe" version
to avoid accidentally removing params which might not be marked as
safe/permitted at this point.)